+++ HISTORY.rst
@@ -2,6 +2,12 @@
 Changelog
 *********
 
+0.2.0 (unreleased)
+------------------
+
+- Add "speed" config option.
+- Fix short option for "--shell".
+
 0.1.0 (2014-06-16)
 ------------------